Improve diagnostic formatting infrastructure
