+++ Sources/SwiftDocC/Infrastructure/Diagnostics/DiagnosticConsumer.swift
@@ -25,13 +25,14 @@ public protocol DiagnosticConsumer: AnyObject {
 }
 
 /// A type that can format received diagnostics in way that's suitable for writing to a destination such as a file or `TextOutputStream`.
+@available(*, deprecated)
 public protocol DiagnosticFormattingConsumer: DiagnosticConsumer {
     /// Options for how problems should be formatted if written to output.
     var formattingOptions: DiagnosticFormattingOptions { get set }
 }
 
 public extension DiagnosticConsumer {
-    // Deprecated for supressing the warning emitted when calling `finalize()`
+    // Deprecated for suppressing the warning emitted when calling `finalize()`
     @available(*, deprecated)
     func flush() throws {
         try finalize()
